refactor(runtime): extract topic logic into its own module
Topic routing, switching, and the `TopicRouter` struct all leave
\`conversation.rs\` for a new \`engine/topic.rs\`. \`conversation.rs\`
drops from 633 to 471 lines and now owns only general lifecycle —
creation, close, persist, title generation. Also suppresses
\`clippy::too_many_arguments\` on \`finalize_run\`; bundling ten args
into a struct would be worse than the warning.
